JDBC with spring using H2 in memory database
So, here we are going to use H2 in memory database for our database connection by using JDBC Template-> it used for 1. Connect with Database, 2. Fire Queries, 3.Process Data, 4. Get Output, so i have taken the H2 dependancy from mvn repository and added to the pom.xml file, my essential files are moved to the model packages , and i created service package for using the service layer with @Service on top of that and i created repository layer package for repo file with the use of @Repository annotation all the objects are connected using @Autowired annotation with using .save ( ) function i request from service layer and i also created the schema.sql and data.sql file in resources package schema.sql consists of creation of table in sql query and data.sql consists of insert query and I also try to add the query from client side which is main class like "Harini" student... so this about entire process , I uploaded each and every file in that