JDBC with Spring using Postgresql database
Follow with H2 database here the simple changes has to be made. All other files remains same just we have to add postgresql dependancy in pom.xml and remove the H2 dependancy and we have to add the url, username, password and driver-class in the Application.properties file which is in resources package. So, that's it we are ready to go we have to open PGadmin4 on our localhost and run the main function in intelliJ it will start to update and insert and create the datas... So, I have attached the essential file alone, complete files->you have to comapre with H2 database