Web Application with Spring
We all know that web applications are run by servlets. What is mean by servlets-> which accepts the request, process the request and give the essential data to the client.
This servlets are managed by  famous server called Tomcat server, which is alight weight, where you can run your servlets.
Hey, we are into spring here, right? why we are taking about servlets?
see, the thing is ultimately, if you want to build the backend, where your'e accepting request from the client, where you need servlets. But, then if you want to use spring, will make sure that you are you are doing in a easy format. It will also help, you to work with servlet behind the scene. We obtain this by spring MVC or spring Web.
Tomcat server:
If you want to run your servlet, you need a container, which is your Tomcat server. Normally, we create a project or web application in java, now if you want to run this on a server, we create a package with the extension (.war) -> web archive, if you are building a console application (.jar) will be used...
.war-> runs on a tomcat server...so, simply you have to create a package (.WAR) of your project and put that on a Tomcat server, so that you can run, by using Tomcat server in our machine. In earlier we have to download the apache Tomcat server in our machine and it consists of webapps folder inside that were you put your project (or) package. 
bin folder-> if you want to start your Tomcat it consists of "startup.sh" and to stop "shutdown.sh" files are there.
Okay!!!!!
What I want is , in my project I want to have Tomcat (InBuild) whether it is possible?
Yes, the concept called "Embedded Tomcat"-> In your project, you can have a embedded tomcat -> you simply, run your project, it'll run your project in that Tomcat.. itself..so, the Advantage is you don't have to do all this configuration.
Now, we are going to create a project, for Servlet understanding, as asual i have taken the spring intializer for spring workspace and in pom.xml file we need 2 dependencies 1. servlet dependency and Embedded Tomcat dependency.
In the main class we have to extends the "HttpServlet" class why means this class mainly works on request and response..
In servlet we have service( ) which consists of mainly two parameters HttpServletRequest and HttpServletResponse
whatever data coming from the client, will be stored in request object AND whatever data you want to send back to the client, will present in your response object, If you want to execute a servlet, you have to send a request from your browser.. in browser-> localhost:8080( default port number ) i have worked on the program and i have uploaded the needed file for your understanding so don't get panic...just relax...
as, we all know that Methods in HTTP:
1.GET-> get data from the server
2.POST-> send data to the server
3.DELETE-> delete something from the server
4.PUT-> update something to the server...
Now we are going to focus on JSP->Jakarta Server Page(Formerly Java Server Page) TO CREATE VIEW TECHNOLOGIES...
Client->servlet->object(->database)->servlet->JSP-> Client will receive response in text/ html page..
Client will send the request, Servlet-> which accpet the request from the client, can do the processing there, you get to the your data through object which connects with database and that data sends to JSP page...
JSP-> is responsible for create that beautiful page[React frontEnd..] which you want...
Servlets->To accept the request from client..
JSP->to show it to the client..
Object->which is the data, which is your object and can access with database also..
NOW, This is were we get a term called MVC->Model, View and Controller
1.Controller[servlets]->Is responsible to accept the request, depends upon, what type of request you get , if it says ok.., I want some data from database, it will take to the database, get the data and may be you wanted some extra processing or something you want to represent the data, you will be using an object(Model) there..
2.Model->so, your object, which you have your data is called a model and then when you want to send this data , the model object  will send it to the client, what happens the controller will send this data to a View(JSP) technology..
3.View-> It will fetch the data, fill up the entire page, and give it back to the client..
SO, that's how MVC works...
*To create a Controller-> you will be using servlets..
*To create a View->You will be using JSP..
*To represent a Model->you will be using simple java class (POJO)...
Finally, you can have a doubt that Tomcat server is a servlet container, only handles the servelets, then how the JSP will run..
Behind the scene-> This JSP page gets converted into Servlets...